// Módulo 4/Atividades/Semana 4/S4-PS01/src/interfaces/user.interface.ts
export interface IUserRequest {
    name: string
    email: string
    password: string
    isAdm: boolean
}

export interface IUser {
    id: string
    name: string
    email: string
    password: string
    isAdm: boolean
    isActive: boolean
    createdAt: Date
    updatedAt: Date
}

export interface IUserWithoutPwd {
    id: string
    name: string
    email: string
    isAdm: boolean
    isActive: boolean
    createdAt: Date
    updatedAt: Date
}

export interface IUserLogin {
    email: string
    password: string
}

export interface IUserUpdate {
    name?: string
    email?: string
    password?: string
}